backend/requirements.txt
flask<2.3
Pillow>=8.0
requests>=2.25
